Support methods on generic types.
Generic type methods are compiled largely in the same way as standalone
generic functions, into a generic function factory. However, unlike a
standalone function, the method factory is registered in the receiver
type factory and is invoked when the type is being instantiated, with
type arguments passed to the generic type. The instantiated method is
then attached to the type instance.